Refactor: move all mouse navigation support code at the same place
